move the continuous distributions into distribution.py
